Remove ExtendedArticles item type. Related to #125
This item type does not seem to be used anywhere, and it's only a more
complicated version of the Articles type. Recent developments made
everything in this type irrelevant or duplicated, so we have no more
reason to keep it around.

I also removed it from the meumobi segment. It was there only for
non-production environments, so it shouldn't affect anything in
production.